Added Turret functionality
